organize the custom property manager to its own file
